requirements.txt
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.7
future==0.18.2
netmiko==3.4.0
ntc-templates==2.1.0
paramiko==2.7.2
psutil==5.8.0
pycparser==2.20
PyNaCl==1.4.0
pyserial==3.5
scp==0.13.5
six==1.16.0
tenacity==7.0.0
textfsm==1.1.2
